# Conflicts:
#	x-pack/plugin/esql-core/src/main/java/org/elasticsearch/xpack/esql/core/type/DataType.java
#	x-pack/plugin/esql/compute/src/main/java/org/elasticsearch/compute/lucene/LuceneQueryEvaluator.java
#	x-pack/plugin/esql/qa/testFixtures/src/main/java/org/elasticsearch/xpack/esql/CsvAssert.java
#	x-pack/plugin/esql/qa/testFixtures/src/main/java/org/elasticsearch/xpack/esql/CsvTestUtils.java
#	x-pack/plugin/esql/qa/testFixtures/src/main/java/org/elasticsearch/xpack/esql/EsqlTestUtils.java
#	x-pack/plugin/esql/qa/testFixtures/src/main/resources/data/dense_vector.csv
#	x-pack/plugin/esql/qa/testFixtures/src/main/resources/dense_vector.csv-spec
#	x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/action/EsqlCapabilities.java
#	x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/expression/function/EsqlFunctionRegistry.java
#	x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/planner/PlannerUtils.java
#	x-pack/plugin/esql/src/test/java/org/elasticsearch/xpack/esql/analysis/AnalyzerTests.java

+++ x-pack/plugin/esql/qa/testFixtures/src/main/java/org/elasticsearch/xpack/esql/CsvAssert.java
@@ -146,8 +146,8 @@ public final class CsvAssert {
                     // Type.asType translates all bytes references into keywords
                     continue;
                 }
-                if (blockType == Type.DOUBLE && expectedType == DENSE_VECTOR) {
-                    // DENSE_VECTOR is internally represented as a double block
+                if (blockType == Type.FLOAT && expectedType == DENSE_VECTOR) {
+                    // DENSE_VECTOR is internally represented as a float block
                     continue;
                 }
                 if (blockType == Type.NULL) {
